Start an experimental LsstHtmlReportExporter
This custom nbconvert exporter is our starting point for creating a
custom ipynb to HTML converter that creates static html versions of LSST
notebook reports.

This is based on the 'full.tpl' template already included in nbconvert,
except now we have a new base template with which to start adding our
own UI layering around the notebook document.

Note that the lsst-report-html entrypoint isn't working yet.