Add publish_instance celery task
This celery task creates the html site from a notebook file and
publishes it to LSST the Docs.

One tricky thing is actually getting the API URL of the edition
corresponding to a slug is tricky. The get_edition_url function solves
this as efficiently as possible, but isn't as good as a single-request
query.

- Adds ltd-conveyor 0.4.0 as a dependency (for S3 uploads)

- Adds configurations for S3 credentials