Integrate publish_instance task with the /notebook
Now the uploaded notebook is passed off to the celery queue and
processed asynchronously by the `publish_instance` task.

When testing the **/notebook route I'm mocking the publish_instance task
so that I don't actually launch it, absent a backend DB.

I wanted to test the content of the arguments passed to publish_instance
but it seems there's variations in the JSON that nbformat is creating.